removed widget tests until i bother to make a real mock of the state management
